python_version/requirements.txt
numpy>=1.21.0
pillow>=8.3.0
deap>=1.3.1